// apps/web/src/api/generated/model/overrideAttendanceSummaryDtoReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type OverrideAttendanceSummaryDtoReason = typeof OverrideAttendanceSummaryDtoReason[keyof typeof OverrideAttendanceSummaryDtoReason];


export const OverrideAttendanceSummaryDtoReason = {
  manual_correction: 'manual_correction',
  policy_exception: 'policy_exception',
  data_fix: 'data_fix',
  reconciliation: 'reconciliation',
} as const;