// apps/web/src/api/generated/model/overrideAttendanceSummaryDtoOverriddenStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type OverrideAttendanceSummaryDtoOverriddenStatus = typeof OverrideAttendanceSummaryDtoOverriddenStatus[keyof typeof OverrideAttendanceSummaryDtoOverriddenStatus];


export const OverrideAttendanceSummaryDtoOverriddenStatus = {
  present: 'present',
  late: 'late',
  early_leave: 'early_leave',
  absent: 'absent',
  leave: 'leave',
  holiday: 'holiday',
  off: 'off',
} as const;